ref: refactor proxy file and login user

+++ src/proxy.ts
@@ -7,37 +7,52 @@ import {
   isAuthRoute,
   UserRole,
 } from './lib/auth-utils';
+import { getUserInfo } from './services/auth/getUserInfo';
 import { deleteCookie, getCookie } from './services/auth/tokenHandler';
+import { getNewAccessToken } from './services/auth/auth.service';
 
 // This function can be marked `async` if using `await` inside
 export async function proxy(request: NextRequest) {
   const pathname = request.nextUrl.pathname;
+  const hasTokenRefreshedParam =
+    request.nextUrl.searchParams.has('tokenRefreshed');
 
-  // const accessToken = request.cookies.get('accessToken')?.value || null;
+  // If coming back after token refresh, remove the param and continue
+  if (hasTokenRefreshedParam) {
+    const url = request.nextUrl.clone();
+    url.searchParams.delete('tokenRefreshed');
+    return NextResponse.redirect(url);
+  }
 
-  const accessToken = await getCookie('accessToken');
+  const tokenRefreshResult = await getNewAccessToken();
+
+  // If token was refreshed, redirect to same page to fetch with new token
+  if (tokenRefreshResult?.tokenRefreshed) {
+    const url = request.nextUrl.clone();
+    url.searchParams.set('tokenRefreshed', 'true');
+    return NextResponse.redirect(url);
+  }
+
+  // const accessToken = request.cookies.get("accessToken")?.value || null;
+
+  const accessToken = (await getCookie('accessToken')) || null;
 
   let userRole: UserRole | null = null;
+  if (accessToken) {
+    const verifiedToken: JwtPayload | string = jwt.verify(
+      accessToken,
+      process.env.JWT_SECRET as string
+    );
 
-  try {
-    if (accessToken) {
-      const verifiedToken: JwtPayload | string = jwt.verify(
-        accessToken,
-        process.env.JWT_SECRET as string
-      );
-
-      if (typeof verifiedToken === 'string') {
-        await deleteCookie('accessToken');
-        await deleteCookie('refreshToken');
-        return NextResponse.redirect(new URL('/login', request.url));
-      }
-
-      userRole = verifiedToken.role;
+    if (typeof verifiedToken === 'string') {
+      await deleteCookie('accessToken');
+      await deleteCookie('refreshToken');
+      return NextResponse.redirect(new URL('/login', request.url));
     }
-  } catch (err) {
-    console.log(err);
-    return NextResponse.redirect(new URL('/login', request.url));
+
+    userRole = verifiedToken.role;
   }
+
   const routerOwner = getRouteOwner(pathname);
   //path = /doctor/appointments => "DOCTOR"
   //path = /my-profile => "COMMON"
@@ -65,12 +80,36 @@ export async function proxy(request: NextRequest) {
     return NextResponse.redirect(loginUrl);
   }
 
-  // Rule 3 : User is trying to access common protected route
+  // Rule 3 : User need password change
+
+  if (accessToken) {
+    const userInfo = await getUserInfo();
+    if (userInfo.needPasswordChange) {
+      if (pathname !== '/reset-password') {
+        const resetPasswordUrl = new URL('/reset-password', request.url);
+        resetPasswordUrl.searchParams.set('redirect', pathname);
+        return NextResponse.redirect(resetPasswordUrl);
+      }
+      return NextResponse.next();
+    }
+
+    if (
+      userInfo &&
+      !userInfo.needPasswordChange &&
+      pathname === '/reset-password'
+    ) {
+      return NextResponse.redirect(
+        new URL(getDefaultDashboardRoute(userRole as UserRole), request.url)
+      );
+    }
+  }
+
+  // Rule 4 : User is trying to access common protected route
   if (routerOwner === 'COMMON') {
     return NextResponse.next();
   }
 
-  // Rule 4 : User is trying to access role based protected route
+  // Rule 5 : User is trying to access role based protected route
   if (
     routerOwner === 'ADMIN' ||
     routerOwner === 'DOCTOR' ||
@@ -82,7 +121,6 @@ export async function proxy(request: NextRequest) {
       );
     }
   }
-  console.log(userRole);
 
   return NextResponse.next();
 }
